+++ rate_control/_buckets/__init__.py
@@ -3,11 +3,9 @@ __all__ = [
     'FixedWindowCounter',
     'LeakyBucket',
     'SlidingWindowLog',
-    'UnlimitedBucket',
 ]
 
 from ._base import Bucket
 from ._fixed_window_counter import FixedWindowCounter
 from ._leaky_bucket import LeakyBucket
 from ._sliding_window_log import SlidingWindowLog
-from ._unlimited import UnlimitedBucket