+++ CHANGELOG.rst
@@ -6,7 +6,9 @@ This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
 Unreleased
 ----------
 
-Nothing changed yet.
+* Now allow passing no bucket argument to the rate controllers.
+
+* **Breaking change**: Removed the now obsolete ``UnlimitedBucket``.
 
 2.0.0
 -----